pyproject.toml
[project]
name = "nanovlm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.0.0",
    "huggingface-hub>=0.34.3",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.54.1",
    "wandb>=0.21.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.12.7",
]
